+++ CHANGES.rst
@@ -4,7 +4,8 @@ CHANGELOG for LaunchKey Python SDK
 3.4.0
 -----
 
-* Added cancel_authorization_request method to service client
+* Added ttl parameter to the DirectoryClient link_device method
+* Added cancel_authorization_request method to the ServiceClient
 
 3.3.0
 -----