+++ CHANGES.rst
@@ -7,6 +7,7 @@ CHANGELOG for LaunchKey Python SDK
 * Added example CLI to codebase
 * Added ttl parameter to the DirectoryClient link_device method
 * Added cancel_authorization_request method to the ServiceClient
+* Fixed bug in ServiceClient which prevented handling of session end webhook request data when presented as a bytearray
 
 3.3.0
 -----